Feat: Rename the types to chatModel
Replace the `manually` to `chat model`
