+++ VSCodeExtension/code-brt/src/services/codeCompletion/types/chatModelCodeComplete.d.ts
